Cosmetic change, updated test cases, added temporary "beaver switch" tests

+++ src/Main.java
@@ -1,6 +1,4 @@
 
-
-
 import parser.*;
 import java.io.*;
 import java.util.*;
@@ -12,38 +10,40 @@ public class Main {
 
 	public static void main(String[] args) {
 		try {
-			if(args.length != 2) {
-				System.err.println("Arguments: SourceFileName DestFileName");
+			if (args.length < 2) {
+				System.err.println("Arguments: SourceFileName DestFileName [--beaver]");
 				System.exit(1);
 			}
 			String source = args[0];
 			String dest = args[1];
+			boolean beaver = false;
+			if (args.length == 3 && args[2].equals("--beaver"))
+				beaver = true;
 			File sourceFile = new File(source);
 			File destFile = new File(dest);
-			if(sourceFile.exists() && destFile.exists()
-			   && sourceFile.lastModified() < destFile.lastModified()) {
+			if (sourceFile.exists() && destFile.exists()
+					&& sourceFile.lastModified() < destFile.lastModified()) {
 				System.out.println("Parser specification " + dest + " need not be regenerated");
-			}
-			else {
-        ASTNode.sourceName = args[0];
+			} else {
+				ASTNode.sourceName = args[0];
 				GrammarScanner scanner = new GrammarScanner(new FileReader(args[0]));
 				GrammarParser parser = new GrammarParser();
 				Object o = parser.parse(scanner);
-				Grammar root = (Grammar)o;
-		ArrayList<String> errors = new ArrayList<String>();
-		ArrayList<String> warnings = new ArrayList<String>();
-        root.errorCheck(errors, warnings);
-        if(!errors.isEmpty()) {
-          System.err.println("There were errors in " + args[0] + ":");
-          for(Iterator iter = errors.iterator(); iter.hasNext(); )
-            System.err.println(iter.next());
-          System.exit(1);
-        }
-        for(Iterator iter = warnings.iterator(); iter.hasNext(); )
-            System.err.println(iter.next());
+				Grammar root = (Grammar) o;
+				ArrayList<String> errors = new ArrayList<String>();
+				ArrayList<String> warnings = new ArrayList<String>();
+				root.errorCheck(errors, warnings);
+				if (!errors.isEmpty()) {
+					System.err.println("There were errors in " + args[0] + ":");
+					for (Iterator iter = errors.iterator(); iter.hasNext(); )
+						System.err.println(iter.next());
+					System.exit(1);
+				}
+				for (Iterator iter = warnings.iterator(); iter.hasNext(); )
+					System.err.println(iter.next());
 				FileOutputStream os = new FileOutputStream(args[1]);
 				PrintStream out = new PrintStream(os);
-				root.pp(out);
+				root.pp(out, beaver);
 				out.close();
 				System.out.println("Parser specification " + dest + " generated from " + source);
 			}
